requirements.txt
python-dotenv==1.0.1
aioconsole==0.7.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
ariadne==0.23.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autoflake==2.2.1
awscli-local==0.22.0
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
executing==2.0.1
fastjsonschema==2.20.0
fqdn==1.5.1
gevent==24.2.1
geventhttpclient==2.3.1
gprof2dot==2022.7.29
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jaconv==0.3.4
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
localstack-client==2.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.0.8
mypy==1.8.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.1.0
Pygments==2.18.0
pykakasi==2.2.1
pyprof2calltree==1.4.5
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.7
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
s3transfer==0.10.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
terminado==0.18.1
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240106
typing_extensions==4.9.0
tzlocal==5.2
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
zope.event==5.0
zope.interface==6.4.post2
